+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+### Breaking Changes
+
+- Moved `ChosenDeploymentAlphaLocatorFlow` and `SessionConnectionFlowInitializer` into the `Fps.Connection` namespace and made them `public`. [#225](https://github.com/spatialos/gdk-for-unity-fps-starter-project/pull/225)
+
+### Internal
+
+- Added tests for the `ChosenDeploymentAlphaLocatorFlow`. [#225](https://github.com/spatialos/gdk-for-unity-fps-starter-project/pull/225)
+
 ## `0.2.6` - 2019-08-05
 
 ### Breaking Changes
@@ -15,8 +23,8 @@
 - Added the `io.improbable.gdk.debug` package as a dependency. [#222](https://github.com/spatialos/gdk-for-unity-fps-starter-project/pull/222)
 
 ### Changed
-- Upgraded to GDK for Unity version `0.2.6`
 
+- Upgraded to GDK for Unity version `0.2.6`
 - The `AdvancedEntityPipeline` now uses the `OwningWorker` component from the PlayerLifecycle feature module [#222](https://github.com/spatialos/gdk-for-unity-fps-starter-project/pull/222)
 - Removed the `DISABLE_REACTIVE_COMPONENTS` scripting define as it is no longer used. [#222](https://github.com/spatialos/gdk-for-unity-fps-starter-project/pull/222)